requirements.txt
absl-py==1.3.0
altair==4.2.0
appnope==0.1.3
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.5
bokeh==2.4.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.0.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter_client==7.4.4
jupyter_core==4.11.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0b4
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scipy==1.7.3
selenium==4.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
streamlit==1.14.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.10.0
